requirements.txt
numpy==1.21.6
tqdm==4.64.1
matplotlib==3.2.2
datasets==2.8.0
transformers==4.25.1
torch==1.13.0
torchvision==0.14.0
notebook==6.5.2
pycocotools
wandb
seaborn
scikit-learn
umap-learn
hdbscan
nltk
sentence-transformers
setuptools
wheel
spacy[cuda-autodetect]
accelerate
diffusers
ujson
swifter
